+++ data/shaders/pbr.vert
@@ -10,7 +10,7 @@ layout (location = 0) in vec3 inPos;
 layout (location = 1) in vec3 inNormal;
 layout (location = 2) in vec2 inUV0;
 layout (location = 3) in vec2 inUV1;
-layout (location = 4) in vec4 inJoint0;
+layout (location = 4) in uvec4 inJoint0;
 layout (location = 5) in vec4 inWeight0;
 layout (location = 6) in vec4 inColor0;
 
@@ -27,7 +27,7 @@ layout (set = 0, binding = 0) uniform UBO
 layout (set = 2, binding = 0) uniform UBONode {
 	mat4 matrix;
 	mat4 jointMatrix[MAX_NUM_JOINTS];
-	float jointCount;
+	uint jointCount;
 } node;
 
 layout (location = 0) out vec3 outWorldPos;
@@ -41,13 +41,13 @@ void main()
 	outColor0 = inColor0;
 
 	vec4 locPos;
-	if (node.jointCount > 0.0) {
+	if (node.jointCount > 0) {
 		// Mesh is skinned
 		mat4 skinMat = 
-			inWeight0.x * node.jointMatrix[int(inJoint0.x)] +
-			inWeight0.y * node.jointMatrix[int(inJoint0.y)] +
-			inWeight0.z * node.jointMatrix[int(inJoint0.z)] +
-			inWeight0.w * node.jointMatrix[int(inJoint0.w)];
+			inWeight0.x * node.jointMatrix[inJoint0.x] +
+			inWeight0.y * node.jointMatrix[inJoint0.y] +
+			inWeight0.z * node.jointMatrix[inJoint0.z] +
+			inWeight0.w * node.jointMatrix[inJoint0.w];
 
 		locPos = ubo.model * node.matrix * skinMat * vec4(inPos, 1.0);
 		outNormal = normalize(transpose(inverse(mat3(ubo.model * node.matrix * skinMat))) * inNormal);
